Fail DecodeHash on trailing parameter junk
Sscanf silently ignores it. Rewrite using a reader so we can make use of
the parse position to detect it.

The errors returned are not exactly the same for the same input as
before. In particular, scanf originated parse errors rather than
ErrInvalidHash are now returned for more inputs.

For completeness, includes tests for leading parameter junk, although
that was not a problem before either.

Could be somewhat faster and allocate less than the previous
implementation, too (no benchmarks run).